ci: consolidate PR-title scripts into a single ci-tools.ts
Replace lint-pr-title.ts, set-pr-labels.ts, and the shared pr-title.ts
module with one ci-tools.ts dispatching on a subcommand argument, so the
title-parsing and label logic live in a single file without a cross-file
import. The workflow now calls `node ./tools/ci-tools.ts <command>`.
