+++ .github/workflows/pull-labeler.yml
@@ -35,7 +35,7 @@ jobs:
       - uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
         with:
           node-version: 24
-      - run: node ./tools/lint-pr-title.ts
+      - run: node ./tools/ci-tools.ts lint-pr-title
         env:
           PR_TITLE: ${{ github.event.pull_request.title }}
 
@@ -55,7 +55,7 @@ jobs:
       - uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
         with:
           node-version: 24
-      - run: node ./tools/set-pr-labels.ts
+      - run: node ./tools/ci-tools.ts set-pr-labels
         env:
           PR_TITLE: ${{ github.event.pull_request.title }}
           PR_NUMBER: ${{ github.event.pull_request.number }}
